+++ docs/Changelog.rst
@@ -11,6 +11,7 @@ Changelog
 - All printing is in red
 - Improved regex to deal with detection of certain version patterns
 - Improved padding of version strings to deal with comparison of certain version patterns
+- Added an option for the user to never be prompted again
 
 **Version: 0.0.1 -- released 24 Mar 2021**
 '''''''''''''''''''''''''''''''''''''''''